Widen Pair::special_lj to double
special_lj was declared int, so the special-bonds factors copied from
force->special_lj for the KOKKOS styles were truncated: any fractional value
became 0 or 1.  Four styles that do not shadow the member with their own
declaration read the truncated values.

+++ src/pair.h
@@ -244,7 +244,7 @@ class Pair : protected Pointers {
 
  protected:
   int instance_me;      // which Pair class instantiation I am
-  int special_lj[4];    // copied from force->special_lj for Kokkos
+  double special_lj[4];    // copied from force->special_lj for Kokkos
 
   // pair_modify settings
   int offset_flag, mix_flag;    // flags for offset and mixing
